requirements.txt
boto3==1.12.29
botocore==1.15.29
click==7.1.1
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
gcloud==0.17.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
